IDPS First Layer/requirements.txt
pandas
numpy
scikit-learn
xgboost
joblib
paho-mqtt
tensorflow
psutil
